+++ src/ui/tab_screen.rs
@@ -9,24 +9,6 @@ use ratatui::{
 use super::ui_legacy::AppUi;
 
 impl AppUi {
-    pub(crate) fn tab_screen_layout() -> Layout {
-        Layout::new(
-            ratatui::layout::Direction::Vertical,
-            vec![
-                Constraint::Fill(1),
-                Constraint::Percentage(85),
-                Constraint::Fill(5),
-            ],
-        )
-    }
-
-    pub(crate) fn work_space_layout() -> Layout {
-        Layout::new(
-            ratatui::layout::Direction::Horizontal,
-            vec![Constraint::Fill(3), Constraint::Fill(1)],
-        )
-    }
-
     pub(crate) fn default_styled_block() -> Block<'static> {
         Block::default()
             .borders(Borders::ALL)
@@ -64,32 +46,6 @@ impl AppUi {
     pub(crate) fn data_popup_rect(work_rect: Rect) -> Rect {
         AppUi::horizontal_equal_layout().split(AppUi::vertical_equal_layout().split(work_rect)[1])
             [1]
-    }
-
-    pub(crate) fn vertical_equal_layout() -> Layout {
-        Layout::vertical(vec![
-            Constraint::Fill(1),
-            Constraint::Fill(1),
-            Constraint::Fill(1),
-        ])
-    }
-    pub(crate) fn vertical_doubled_layout() -> Layout {
-        Layout::vertical(vec![
-            Constraint::Fill(1),
-            Constraint::Fill(2),
-            Constraint::Fill(1),
-        ])
-    }
-    pub(crate) fn horizontal_equal_layout() -> Layout {
-        Layout::horizontal(vec![
-            Constraint::Fill(1),
-            Constraint::Fill(1),
-            Constraint::Fill(1),
-        ])
-    }
-
-    pub(crate) fn vertical_double_popup_layout() -> Layout {
-        Layout::vertical(vec![Constraint::Fill(1), Constraint::Fill(1)])
     }
 
     pub(crate) fn edit_path_active_block() -> Block<'static> {
